pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hiscanner"
version = "0.2.0"
description = "Single-cell Copy Number Variation Analysis Pipeline"
readme = "README.md"
authors = [{name = "Yifan Zhao", email = "yifnzhzo@gmail.com"}]
license = {file = "LICENSE"}
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.20.0",
    "pandas>=1.3.0",
    "scipy>=1.7.0",
    "scikit-learn>=1.0.0",
    "hmmlearn>=0.2.7",
    "pysam>=0.16.0",
    "matplotlib>=3.4.0",
    "seaborn>=0.11.0",
    "snakemake>=7.0.0",
    "pyyaml>=5.4.0",
    "click>=8.0.0",
    "colorlog>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
    "black>=22.0",
    "flake8>=3.9",
    "mypy>=0.910",
    "isort>=5.9",
]

[project.scripts]
hiscanner = "hiscanner.cli:main"

[tool.setuptools]
packages = ["hiscanner"]

[tool.setuptools.package-data]
hiscanner = ["resources/*"]